Refactoring audit to be more modular and pluggable
