+++ src/client/app/admin/user-management/audit/user-audit.component.ts
@@ -1,0 +1,10 @@
+import { AuditObjectTypes } from '../../../audit/audit.classes';
+import { DefaultAudit } from '../../../audit/audit-object.component';
+import { Component } from '@angular/core';
+
+@Component({
+	selector: 'user',
+	templateUrl: './user.audit.view.html'
+})
+export class UserAudit extends DefaultAudit {}
+AuditObjectTypes.registerType('user', UserAudit);
