+++ src/client/app/messages/messages.module.ts
@@ -5,16 +5,22 @@ import { SocketService } from '../core/socket.service';
 import { MessagesComponent } from './messages.component';
 import { UtilModule } from '../shared/util.module';
 import { CommonModule } from '@angular/common';
+import { MessageAudit } from './audit/message-audit.component';
 
 @NgModule({
 	imports: [
 		CommonModule,
 		UtilModule
 	],
+	entryComponents: [
+		MessageAudit
+	],
 	exports: [
+		MessageAudit,
 		MessagesComponent
 	],
 	declarations: [
+		MessageAudit,
 		MessagesComponent
 	],
 	providers: [
